Add description field and change .equals to .equalsIgnoreCase

+++ src/main/java/com/conorthedev/sileogenerator/SileoGenerator.java
@@ -49,34 +49,34 @@ public class SileoGenerator {
             /*
                 If the first argument is equal to 'create' start the JSON creation, first gathering info from the user
              */
-            if(args[0].equals("create")) {
+            if(args[0].equalsIgnoreCase("create")) {
                 // Start a new reader instance
                 Scanner reader = new Scanner(System.in);
 
                 // Get the package name
                 System.out.println("Package Name: ");
-                name = reader.next();
+                name = reader.nextLine();
 
                 // Get the header URL
                 System.out.println("Header Image URL (type none for none): ");
-                headerUrl = reader.next();
+                headerUrl = reader.nextLine();
 
                 // Get the tint colour
                 System.out.println("Tint Colour (hexadecimal, e.g: #f75454): ");
-                tintColour = reader.next();
+                tintColour = reader.nextLine();
 
                 // Get the package description
                 System.out.println("Short description of your package: ");
-                description = reader.next();
+                description = reader.nextLine();
 
                 // Display the gathered information to the user for confirmation
-                System.out.println("\nPackage Information:\nName: " + name + "\nHeader Image URL: " + headerUrl + "\nTint Colour: " + tintColour);
+                System.out.println("\nPackage Information:\nName: " + name + "\nHeader Image URL: " + headerUrl + "\nTint Colour: " + tintColour + "\nDescription: " + description);
 
                 // Ask the user if this information is correct
                 System.out.println("Is this information correct? (y/n)");
-                String isOk = reader.next();
+                String isOk = reader.nextLine();
 
-                if(isOk.equals("y")) {
+                if(isOk.equalsIgnoreCase("y")) {
                     // The information is correct so we start generating a depiction
                     System.out.println("\nGenerating Sileo Depiction... This may take a few moments...");
 
@@ -86,7 +86,7 @@ public class SileoGenerator {
                     // Put the user's package info into the JSONObject (generatedJson)
                     generatedJson.put("minVersion", "0.1");
                     // If the headerURL isn't blank put it in, otherwise skip it
-                    if (!headerUrl.equals("none")) {
+                    if (!headerUrl.equalsIgnoreCase("none")) {
                         generatedJson.put("headerImage", headerUrl);
                     }
                     generatedJson.put("tintColor", tintColour);
@@ -126,22 +126,46 @@ public class SileoGenerator {
 
                     // Write the json to a file with the package name at the root of the directory
                     File json = new File(runDirectory, name + ".json");
-                    try {
-                        if(json.createNewFile()) {
-                            // The file was successfully created, now write to it
+                    if(!json.exists()) {
+                        try {
+                            if(json.createNewFile()) {
+                                // The file was successfully created, now write to it
 
-                            Writer output = new BufferedWriter(new FileWriter(json));
-                            output.write(generatedJsonOutput);
-                            output.close();
+                                Writer output = new BufferedWriter(new FileWriter(json));
+                                output.write(generatedJsonOutput);
+                                output.close();
 
-                            System.out.println("Wrote JSON file to: " + json.getAbsolutePath());
-                        } else {
-                            // The file was unsuccessfully created
-                            System.out.println("Error writing JSON file to: " + json.getAbsolutePath());
+                                System.out.println("Wrote JSON file to: " + json.getAbsolutePath());
+                            } else {
+                                // The file was unsuccessfully created
+                                System.out.println("Error writing JSON file to: " + json.getAbsolutePath());
+                            }
+                        } catch (IOException e) {
+                            // An error occured, oh noes!
+                            e.printStackTrace();
                         }
-                    } catch (IOException e) {
-                        // An error occured, oh noes!
-                        e.printStackTrace();
+                    } else {
+                        // The file already exists, ask the user if we should overwrite it
+                        System.out.println("The file: + " + json.getAbsolutePath() + " already exists, overwrite? (y/n)");
+                        String overwrite = reader.nextLine();
+
+                        if(overwrite.equalsIgnoreCase("y")) {
+                            // We have permission to overwrite the file
+                            Writer output = null;
+                            try {
+                                output = new BufferedWriter(new FileWriter(json));
+                                output.write(generatedJsonOutput);
+                                output.close();
+                            } catch (IOException e) {
+                                e.printStackTrace();
+                            }
+                        } else {
+                            // Close the program
+                            System.out.println("Exiting program...");
+                            System.exit(0);
+                        }
+
+                        System.out.println("Wrote JSON file to: " + json.getAbsolutePath());
                     }
                 } else {
                     // The information is not correct so we exit the program
